main page skeleton done and added router functionality
