--- client/src/pages/About.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function About() {
-    return (
-        <div>
-            
-        </div>
-    )
-}
-
-export default About;
