+++ client/src/pages/About/About.js
@@ -1,0 +1,40 @@
+import React from "react";
+import "./style.css"
+
+function About() {
+    return (
+        //Background, placeholder
+        <div className="view">
+
+            <div className="container">
+                <div className="about-text">
+                    <h1>
+
+                    </h1>
+                    <p>
+
+                    </p>
+                </div>
+
+                <div className="diana">
+
+                </div>
+                <div className="don">
+
+                </div>
+                <div className="joanna">
+
+                </div>
+                <div className="elodie">
+
+                </div>
+                <div className="james">
+
+                </div>
+            </div>
+
+        </div>
+    )
+}
+
+export default About;
