+++ client/src/pages/About/style.css
@@ -1,0 +1,6 @@
+.view {
+    background-image: url('https://images.pexels.com/photos/281260/pexels-photo-281260.jpeg?auto=compress&cs=tinysrgb&dpr=3&h=750&w=1260');
+    background-repeat: no-repeat;
+    height: auto;
+    width: auto;
+}
