--- client/src/pages/Main.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function Main() {
-    return (
-        <div>
-            
-        </div>
-    )
-}
-
-export default Main;
